Update compile fail tests to use usize.

+++ src/test/compile-fail/assign-to-method.rs
@@ -9,7 +9,7 @@
 // except according to those terms.
 
 struct cat {
-  meows : uint,
+  meows : usize,
 
   how_hungry : isize,
 }
@@ -18,7 +18,7 @@ impl cat {
     pub fn speak(&self) { self.meows += 1u; }
 }
 
-fn cat(in_x : uint, in_y : isize) -> cat {
+fn cat(in_x : usize, in_y : isize) -> cat {
     cat {
         meows: in_x,
         how_hungry: in_y
