+++ src/test/compile-fail/issue-17718-patterns.rs
@@ -8,9 +8,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-static A1: uint = 1;
-static mut A2: uint = 1;
-const A3: uint = 1;
+static A1: usize = 1;
+static mut A2: usize = 1;
+const A3: usize = 1;
 
 fn main() {
     match 1u {
