+++ src/test/compile-fail/regions-addr-of-upvar-self.rs
@@ -8,16 +8,16 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use std::uint;
+use std::usize;
 
 struct dog {
-    food: uint,
+    food: usize,
 }
 
 impl dog {
     pub fn chase_cat(&mut self) {
         let _f = |&:| {
-            let p: &'static mut uint = &mut self.food; //~ ERROR cannot infer
+            let p: &'static mut usize = &mut self.food; //~ ERROR cannot infer
             *p = 3u;
         };
     }
